解决get_folder获取不到根目录 添加下载css

+++ html_down/functions.rb
@@ -10,6 +10,9 @@ end
 
 # 下载文件
 def download(url, filename)
+	if File.exist? filename then
+		return
+	end
 	puts "download: #{url}"
 	data=open(url) { |f| f.read }
 	open(filename, "wb"){ |f| f.write data }
@@ -18,7 +21,7 @@ end
 
 def get_script_src(content)
 	arr = []
-	content.scan(%r{<script.*src=["|](.*)["|'].*></script>}).each do |item|
+	content.scan(%r{<script.*src=["|](.*?)["|'].*></script>}).each do |item|
 		arr.push(item[0])
 	end
 	arr
@@ -27,13 +30,28 @@ def get_script_src(content)
 #	end
 end
 
+def get_css_src(content)
+	arr = []
+	content.scan(%r{<link.*href=["|](.*?)["|'].*>}).each do |item|
+		arr.push(item[0])
+	end
+	arr
+end
+
+
 def get_folder(url)
+	puts "url:" + url
 	lindex = url.index("/")+1
 	if url[lindex+1] == "/" then
 		lindex += 1
 	end
 	rindex = url.rindex("/")-lindex
-	url[lindex, rindex]
+	folder = url[lindex, rindex]
+	# 当位于根目录的时候返回""
+	if folder==nil then
+		folder = ""
+	end
+	folder
 end
